# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": " Mojo Api",
        "description": "MOJO API Descrição",
        "contact": {
            "email": "wesleymotabr@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://back.mojosolutions.co/api/",
            "description": "Server Mojo"
        }
    ],
    "paths": {
...
        "/events/{id}": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Get an Event",
                "description": "Returns list of Events",
                "operationId": "single_event",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "put": {
                "tags": [
                    "Events"
                ],
                "summary": "Edit an Event",
                "description": "Returns list of Events",
                "operationId": "update_event",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "title",
                                    "code",
                                    "category",
                                    "start_time",
                                    "end_time",
                                    "location",
                                    "description",
                                    "img",
                                    "value",
                                    "recurrence"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "integer"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "code": {
                                        "type": "string"
                                    },
                                    "category": {
                                        "type": "string"
                                    },
                                    "start_time": {
                                        "type": "datetime"
                                    },
                                    "end_time": {
                                        "type": "datetime"
                                    },
                                    "location": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "img": {
                                        "type": "string"
                                    },
                                    "value": {
                                        "type": "double"
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "tags": {
                                        "type": "string"
                                    },
                                    "recurrence": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "type": 1,
                                    "title": "olg",
                                    "code": "#a101",
                                    "category": "1",
                                    "start_time": "2019-12-28 20:45:12",
                                    "end_time": "2019-08-30 23:05:12",
                                    "location": "Carrin",
                                    "description": "legal",
                                    "img": null,
                                    "value": 10,
                                    "user_id": 1,
                                    "tags": "teste, feliz, sad, test",
                                    "recurrence": "monthly"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Events"
                ],
                "summary": "Delete an Event",
                "description": "Returns list of Events",
                "operationId": "delete_event",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
...
        "/tags": {
            "get": {
                "tags": [
                    "tags"
                ],
                "summary": "Get list of tags",
                "description": "Returns list of tags",
                "operationId": "gettags",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "tags"
                ],
                "summary": "Add an tag",
                "description": "Returns list of tags",
                "operationId": "add_tag",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/tags/{id}": {
            "get": {
                "tags": [
                    "tags"
                ],
                "summary": "Get an tag",
                "description": "Returns list of tags",
                "operationId": "single_tag",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "put": {
                "tags": [
                    "tags"
                ],
                "summary": "Edit an tag",
                "description": "Returns list of tags",
                "operationId": "update_tag",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "delete": {
                "tags": [
                    "tags"
                ],
                "summary": "Delete an tag",
                "description": "Returns list of tags",
                "operationId": "delete_tag",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
...
    },
...
}